requirements.txt
scikit-learn>=1.3
seqeval>=1.2.2
krippendorff>=0.7
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
shap>=0.43
gradio>=4.0
openai>=1.0
